Revert "[PCCP-3686] [Dev][Plugin Issue] Port missing Network Pool Provider from embedded"
