+++ src/main/groovy/com/morpheusdata/scvmm/sync/IpPoolsSync.groovy
@@ -1,6 +1,5 @@
 package com.morpheusdata.scvmm.sync
 
-import com.morpheusdata.model.NetworkPoolServer
 import com.morpheusdata.scvmm.ScvmmApiService
 import com.morpheusdata.core.MorpheusContext
 import com.morpheusdata.core.data.DataFilter
@@ -26,13 +25,11 @@ class IpPoolsSync {
     private Cloud cloud
     private ScvmmApiService apiService
     private LogInterface log = LogWrapper.instance
-    private NetworkPoolServer networkPoolServer
 
-    IpPoolsSync(MorpheusContext morpheusContext, Cloud cloud, NetworkPoolServer server) {
+    IpPoolsSync(MorpheusContext morpheusContext, Cloud cloud) {
         this.cloud = cloud
         this.morpheusContext = morpheusContext
         this.apiService = new ScvmmApiService(morpheusContext)
-        this.networkPoolServer = server
     }
 
     def execute() {
@@ -56,7 +53,7 @@ class IpPoolsSync {
             def listResults = apiService.listNetworkIPPools(scvmmOpts)
 
             if (listResults.success == true) {
-                def poolType = new NetworkPoolType(code: 'scvmm-plugin-ipam')
+                def poolType = new NetworkPoolType(code: 'scvmm')
                 def objList = listResults.ipPools
                 def networkMapping = listResults.networkMapping
 
@@ -127,8 +124,6 @@ class IpPoolsSync {
                     poolRangeAdds << newRange
                     add.addToIpRanges(newRange)
                 }
-
-                morpheusContext.async.network.pool.create(this.networkPoolServer.id, [add]).blockingGet()
             }
 
             if(networkPoolAdds.size() > 0){
@@ -309,13 +304,6 @@ class IpPoolsSync {
 
                     if(existingItem.subnetAddress != info.networkAddress) {
                         existingItem.subnetAddress = info.networkAddress
-                        doSave = true
-                    }
-
-                    if (existingItem?.poolServer?.id != this.networkPoolServer.id || existingItem.parentId != this.networkPoolServer.id) {
-                        existingItem.poolServer = this.networkPoolServer
-                        existingItem.parentId = this.networkPoolServer.id
-                        existingItem.parentType = "NetworkPoolServer"
                         doSave = true
                     }
